+++ ps2/Solution/src/ISort.java
@@ -3,10 +3,10 @@ import java.util.stream.Collectors;
 
 public interface ISort {
 
-     static Map<Integer, Integer> sortByValueReverseOrder(final Map<Integer, Integer> wordCounts) {
+     static Map<Integer, Double> sortByValueReverseOrder(final Map<Integer, Double> wordCounts) {
         return wordCounts.entrySet()
                 .stream()
-                .sorted((Map.Entry.<Integer, Integer>comparingByValue().reversed()))
+                .sorted((Map.Entry.<Integer, Double>comparingByValue().reversed()))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (e1, e2) -> e1, LinkedHashMap::new));
     }
